Add TrapMenu, WidgetChil8, WidgetGroup script, some voxel obj flags,misc.

+++ include/PlayerData.h
@@ -183,7 +183,7 @@ void ProcessPlayers(PlayerHandler *,s16);
 void FUN_80017330(playerData *,float,float,float,float);
 void FUN_80017388(playerData *,float);
 void some_player_render_sub(playerData *,SceneData *,vec3f*,u8,u16);
-void set_sun_light(SceneData *,u32,voxelObject*,u8);
+void set_sun_light(SceneData *,u16,voxelObject*,u8);
 Gfx * renderPlayers(PlayerHandler *,Gfx *,short,short,short);
 void edit_playerdat_combat_pos(playerData *,vec3f *);
 void FUN_800187f4(attachmentNode *);
